Blank Line Cleanup
Cleaned up empty lines to remove spaces. Editor automatically removes these
spaces when saving.

+++ octree/include/node.h
@@ -46,7 +46,7 @@ class node
             }
             children[childNumber] = new node<type>(this,tempOrigin,tempHalfDimension,level,maxMembers,maxTheta);
     }
-    
+
     void pushMember(const member<type> &member)
     {
         // Checks if child is already created. If not, creates the right child before adding member.
@@ -61,12 +61,12 @@ class node
             children[child]->addMember(member);
         }
     }
-    
+
     void addChild(node<type>* child, int childNumber)
     {
         children[childNumber] = child;
     }
-    
+
     void addLevel()
     {
         //Recursively adds one to the level of each node if parent is added
@@ -81,7 +81,7 @@ class node
             }
         }
     }
-    
+
     std::vector<type*> membersToObjects()
     {
         std::vector<type*> objects;
@@ -92,7 +92,7 @@ class node
         }
         return objects;
     }
-    
+
 public:
     node(node<type>* parent_ptr,Eigen::Vector3d origin,Eigen::Vector3d halfDimension, short parent_level,  short maxMembers, double maxTheta)
       : parent(parent_ptr),origin(origin),halfDimension(halfDimension),maxMembers(maxMembers),maxTheta(maxTheta),multExp(nullptr)
@@ -101,7 +101,7 @@ public:
         {
             children[i] = NULL;
         }
-        
+
         level = parent_level+1;
     }
 
@@ -115,7 +115,7 @@ public:
             children[i] = new node<type>(*copy.children[i]);
         }
     }
-    
+
     ~node()
     {
         if (!isLeafNode())
@@ -129,7 +129,7 @@ public:
             }
         }
     }
-    
+
     node<type> operator=(node<type> other)
     {
         parent = other.parent;
@@ -144,13 +144,13 @@ public:
         }
         return this;
     }
-    
-    
+
+
     void setMaxMembers(const int &max)
     {
         maxMembers = max;
     }
-    
+
     void createParent(int corner)
     {
         // Corner   0 1 2 3 4 5 6 7
@@ -167,7 +167,7 @@ public:
             tempHalfDimension[i] = 2*halfDimension[i];
         }
         parent = new node<type>(NULL,tempOrigin,tempHalfDimension,level-1,maxMembers,maxTheta);
-        
+
         int child = 0;
         for (int i=0; i<3; i++)
         {
@@ -176,7 +176,7 @@ public:
         parent->addChild(this,child); //Adds current node to parent
         addLevel();
     }
-    
+
     void addMember(const member<type> &member)
     {
         if (isLeafNode())  //If no children exist, add member to current node.  If they do, add member to proper child.
@@ -196,7 +196,7 @@ public:
             pushMember(member);
         }
     }
-    
+
     bool isLeafNode()
     {
         bool flag = true;
@@ -210,15 +210,15 @@ public:
         }
         return flag;
     }
-    
+
     bool isSameLevel(node<type>* otherNode)
     {
         bool isSame = this->getLevel()==otherNode->getLevel()?true:false;
         return isSame;
     }
-    
+
     Eigen::Vector3d calcVel(Eigen::Vector3d POI){
-        
+
         Eigen::Vector3d velInfl = Eigen::Vector3d::Zero();
         if(this->isLeafNode())
         {
@@ -243,12 +243,12 @@ public:
                 }
             }
         }
-        
+
         return velInfl;
     }
-    
+
     Eigen::Vector3d calcVel(type* part){
-        
+
         Eigen::Vector3d velInfl = Eigen::Vector3d::Zero();
         if(this->isLeafNode())
         {
@@ -276,12 +276,12 @@ public:
                 }
             }
         }
-        
+
         return velInfl;
     }
-    
+
     Eigen::Vector3d calcStretch(type* part){
-        
+
         Eigen::Vector3d stretchInfl = Eigen::Vector3d::Zero();
         if(this->isLeafNode())
         {
@@ -309,18 +309,18 @@ public:
                 }
             }
         }
-        
+
         return stretchInfl;
     }
 
-    
+
     Eigen::Vector3d calcDiff(type* part){
-        
+
         Eigen::Vector3d diffInfl = Eigen::Vector3d::Zero();
         if(this->isLeafNode())
         {
             std::vector<type*> nParts = this->getMembers();
-            for(int i=0; i<nParts.size(); i++)
+            for(size_t i=0; i<nParts.size(); i++)
             {
                 if(nParts[i] != part) // Kroneger delta func.
                 {
@@ -343,12 +343,12 @@ public:
                 }
             }
         }
-        
+
         return diffInfl;
     }
 
-    
-    
+
+
 //    bool isFarField(node<type>* otherNode)
 //    {
 //        if(otherNode->getLevel() == 2){
@@ -356,23 +356,23 @@ public:
 //        }
 //        double centerToCenterDist = (this->getOrigin()-otherNode->getOrigin()).norm();
 //        bool sameLevel = isSameLevel(otherNode);
-//        
+//
 //        if(std::abs(centerToCenterDist) > 2*(this->getHalfDimension()).norm() && sameLevel){
 //            return true;
 //        }else{
 //            return false;
 //        }
 //    };
-    
+
     bool isFarField(node<type>* otherNode){
         // Follows documentation in FTM chapter of Connor's thesis
 
         // Width of Node
         double s = 2*this->halfDimension.x(); // The 'x' can be used because the implementation of the octree keeps each node cubic
-        
+
         // Particle distance
         double d = (this->multExp->pos - otherNode->multExp->pos).norm();
-        
+
         // theta = s/d;
         if(s/d < maxTheta)
             // maxTheta was put in the particle class to allow for different values for panels if BH is implemented in the future
@@ -384,15 +384,15 @@ public:
             return false;
         }
     };
-    
+
     bool isFarField(Eigen::Vector3d POI){
-        
+
         // Width of Node
         double s = 2*this->halfDimension.x(); // The 'x' can be used because the implementation of the octree keeps each node cubic
-        
+
         // Particle distance
         double d = (this->multExp->pos - POI).norm();
-        
+
         // theta = s/d;
         if(s/d < maxTheta)
         {
@@ -403,13 +403,13 @@ public:
             return false;
         }
     };
-    
-    
+
+
     int getChildContainingMember(const member<type> &member)
     {
         return getChildContainingPnt(member.getRefPoint());
     }
-    
+
     int getChildContainingPnt(const Eigen::Vector3d &pnt)
     { // Finds child quadrant
         int child = 0;
@@ -425,10 +425,10 @@ public:
         {
             child |= 1;
         }
-        
+
         return child;
     }
-    
+
     std::vector<type*> getMembers()
     {
         if (isLeafNode())
@@ -451,9 +451,9 @@ public:
             }
             return recursiveMembers;
         }
-        
+
     }
-    
+
     std::vector<node<type>*> getSubNodes()
     {
         if (isLeafNode())
@@ -480,7 +480,7 @@ public:
             return recursiveNodes;
         }
     }
-    
+
     std::vector<type*> getMembers(node<type>* exception)
     {
         // Used if searching tree starting at bottom to avoid searching the same node twice.
@@ -510,7 +510,7 @@ public:
             return recursiveMembers;
         }
     }
-    
+
     bool hasChildren()
     {
         for(int i=0; i<8; i++)
@@ -523,7 +523,7 @@ public:
         }
         return false;
     }
-    
+
     std::vector<node<type>*> getChildren()
     {
         std::vector<node<type>*> children;
@@ -537,23 +537,23 @@ public:
         }
         return children;
     }
-    
+
     std::vector<type*> getChildExpans()
     {
         std::vector<node<type>*> children = this->getChildren();
         std::vector<type*> childExps;
-        
+
         for(int i=0; i<children.size(); i++)
         {
             childExps.push_back(children[i]->multExp);
         }
-        
+
         return childExps;
     }
-    
+
     type* multExp; // Public because will access a lot
     void setMultExp(type* exp) {multExp = exp;};
-    
+
     Eigen::Vector3d getOrigin() {return origin;}
     Eigen::Vector3d getHalfDimension() {return halfDimension;}
     node<type>* getChild(int childNumber) {return children[childNumber];}
